requirements.txt
# A2A World Platform - Python Dependencies
# Core backend and agent system requirements

# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database and ORM
asyncpg==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.13.1
geoalchemy2==0.14.2

# Redis client
redis==5.0.1
aioredis==2.0.1

# NATS messaging
nats-py==2.6.0

# Consul service discovery
python-consul==1.1.0

# Geospatial libraries
shapely==2.0.2
geopandas==0.14.1
fiona==1.9.5
pyproj==3.6.1
rasterio==1.3.9

# KML/GeoJSON parsing
fastkml==0.12
geojson==3.1.0

# Machine Learning and AI
scikit-learn==1.3.2
hdbscan==0.8.33
numpy==1.25.2
pandas==2.1.4
scipy==1.11.4

# Text processing and embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2
ollama==0.1.7

# HTTP clients and utilities
httpx==0.25.2
aiofiles==23.2.1
python-multipart==0.0.6

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Environment and configuration
python-dotenv==1.0.0
pyyaml==6.0.1

# Logging and monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Code quality
flake8==6.1.0
black==23.12.0
isort==5.13.2

# Development utilities
watchdog==3.0.0
python-dateutil==2.8.2
click==8.1.7

# Email functionality
aiosmtplib==3.0.1
jinja2==3.1.2

# File processing
pillow==10.1.0
openpyxl==3.1.2

# Async utilities
anyio==4.2.0

# Production server
gunicorn==21.2.0

# Statistics and validation
statsmodels==0.14.1
pingouin==0.5.4

# Clustering algorithms
scikit-learn-extra==0.3.0

# Graph processing (for pattern relationships)
networkx==3.2.1

# Time series analysis
pytz==2023.3.post1

# Memory profiling and optimization
memory-profiler==0.61.0

# Async database utilities
databases[postgresql]==0.8.0

# Configuration validation
cerberus==1.3.5

# UUID utilities
uuid==1.30